There are some changes from API level 28.
From API level 28 new security layer is added that it won't load the resources from the https. We manually have to specify that it is ok to load the resources from http.

So, Glide won't be able to load the network resource easily.

So We have added new network_security_config.xml which contains the rule and allowed sub-domains.

From AndroidManifest.xml we are referring the networkSecurityConfig file.