Added the functionality of RecyclerView. Which contains the adapter, interface for eventListeners and separate viewHolder classes.
Here we have created the separate ViewHolder class instead creating it inside the adapter class, because of the following reasons.
In this application there will be one ui for the category list,
another for loading,
another for error,
another for recipe

and all will be the part of recyclerView. so, instead of different-different recyclerViews we will be going to create different-different ViewHolders and changing it to the recyclerView.

So for the easy readability we have created the separate ViewHolder Class.