Added the feature - "default category list of recipes" in the recyclerview. For that, I have created new viewHolder for the recipe categories called CategoryViewHolder.java . All the recipe categories and their correspondence images are statically defined in the Constants.java file.
In the Adapter class, created new type to identify the CategoryViewHolder.

From the RecipeListActivity.java, displaySearchCategories() of adapter class is called. Which will take the categories from the Constants.java and builds a list of the recipes. And in those recipes following values have been set:
title,
image,
prepareTime = -1

Here, prepareTime = -1, because to identify that the item in the recipes List is a category and not the recipe. And when the notifyDataSetChanged() will be called, it will rebuild the adapter. And from there the itemType will be identified. And the CategoryViewHolder layout will be displayed.

And finally modified the onCategoryClick() in the RecipeListActivity.java So that, when we click on the category it will show the progress bar and make a search query for that selectd category and displays the result of all the recipes.