Added the functionality of pagination. It means loading the more results whenever requested (Here, whenever the user reaches to the end of the recipe list in recycler view at that time).
For that I have used the scrollListener of the recyclerView and there, checking whether the recycler view is at the bottom or I can say that recyclerView is displaying its last item on the screen. And at this time calling the searchNextRecords() method of the viewModel.

From viewModel again calling the searchNextRecords() method of the repository.

In the repository I will save the details of query and diet when that request made at the first time and when the searchNextRecords() will be called from the viewModel I am passing those stored query, diet and also passing the recordsToSkip to the recipeSearchApi().

And from there call will be made to the RecipeApiClient.java and there the response will be retrieved and appended to the existing list.

To check how many records do I need to skip, for that whenever the response will be retrieved in the RecipeApiClient.java at that time storing the number of records in a variable. And that variable is used in the repository to get the number of records to skip and passing it along with the query and diet to the searchRecipeApi().

Meanwhile, when this process is running in the background, the loading layout will be added to the recycler View and it will be displaying as the last item.