Added the complete functionality of displaying the individual recipe with details when the user selects particular recipe from the list.
First created the activity called RecipeActivity and its correspondence class and layout file.

Also added the functionality to make the request (calling the individual recipe api) and get the response (in the form of RecipeResponse.java) to the RecipeApiClient.java file with the custom Runnable interface.

And from the RecipeApiClient - recipeResponse variable is observed along the way to the RecipeActivity.java file.

After that in this activity first getting the id of the recipe which is selected by the user. Passing it to the chain:

RecipeActivity -> viewModel -> repository -> RecipeApiClient = response
response = RecipeApiClient <- repository <- viewModel <- RecipeActivity

Here created separate ViewModel for the activity for the simplicity.