Added the functionality of query exhausted.
When there are no more results available it will show the query exhausted (no more results...) message at the end of the list.

For this functionality, I have added a MediatorLiveData in the repository, which is used to get the data from the ApiClient and modify it and then it will deliver that data to the viewModel.