Added the functionality of switching between the multiple api keys. If the daily limit of one API key is exceeded then it will show one popup message and ask the user to restart the application. In between the api key will be switched.
For this functionality created ApiKeyUtility.java which will store the current API key to the shared preferences. And different methods to set the current API keys from the list of API keys.

Utility.java file contains a method which will show the alert box to the user. And also from there the new API key will be switched.

Here, to check whether the API quota is over or not - I have kept a MutableLiveData boolean variable in the RecipeApiClient.java and from there its value will be changed.
And observing that along the way to the both Activities.

In Constants.java file, created one array of api keys.

Also added the new API Keys.